@@ -34,42 +34,42 @@
   },
   "homepage": "https://github.com/thehackerati/react-app-template#readme",
   "dependencies": {
-    "express": "^4.13.4",
+    "express": "^4.15.3",
     "immutable": "^3.8.1",
-    "react": "^15.1.0",
-    "react-dom": "^15.0.1",
-    "react-redux": "^4.4.5",
-    "redux": "^3.5.2",
-    "redux-thunk": "^2.0.1"
+    "react": "^15.6.1",
+    "react-dom": "^15.6.1",
+    "react-redux": "^5.0.5",
+    "redux": "^3.7.1",
+    "redux-thunk": "^2.2.0"
   },
   "devDependencies": {
-    "babel-cli": "^6.9.0",
-    "babel-core": "^6.3.17",
-    "babel-eslint": "^6.1.2",
-    "babel-loader": "^6.2.0",
-    "babel-preset-es2015": "^6.1.3",
-    "babel-preset-react": "6.3.13",
-    "babel-preset-stage-0": "^6.3.13",
-    "chai": "^3.5.0",
-    "chai-immutable": "^1.5.4",
-    "enzyme": "^2.3.0",
-    "eslint": "^3.1.1",
-    "eslint-plugin-react": "^5.2.2",
-    "extract-text-webpack-plugin": "^1.0.1",
-    "html-webpack-plugin": "^2.19.0",
-    "jsdom": "^9.2.1",
-    "mocha": "^2.5.3",
-    "node-libs-browser": "^0.5.2",
-    "react-addons-test-utils": "^15.1.0",
-    "react-hot-loader": "^3.0.0-alpha.8",
-    "reactotron": "^0.6.1",
-    "redux-devtools": "^3.2.0",
-    "redux-devtools-dock-monitor": "^1.1.1",
-    "redux-devtools-log-monitor": "^1.0.11",
-    "sinon": "^1.17.4",
-    "stats-webpack-plugin": "^0.3.1",
-    "webpack": "^1.12.2",
-    "webpack-dev-middleware": "^1.6.1",
-    "webpack-hot-middleware": "^2.10.0"
+    "babel-cli": "^6.24.1",
+    "babel-core": "^6.25.0",
+    "babel-eslint": "^7.2.3",
+    "babel-loader": "^7.1.0",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
+    "babel-preset-stage-0": "^6.24.1",
+    "chai": "^4.0.2",
+    "chai-immutable": "^1.6.0",
+    "enzyme": "^2.9.1",
+    "eslint": "^4.1.1",
+    "eslint-plugin-react": "^7.1.0",
+    "extract-text-webpack-plugin": "^2.1.2",
+    "html-webpack-plugin": "^2.29.0",
+    "jsdom": "^11.0.0",
+    "mocha": "^3.4.2",
+    "node-libs-browser": "^2.0.0",
+    "react-addons-test-utils": "^15.6.0",
+    "react-hot-loader": "^1.3.1",
+    "reactotron": "^0.9.0",
+    "redux-devtools": "^3.4.0",
+    "redux-devtools-dock-monitor": "^1.1.2",
+    "redux-devtools-log-monitor": "^1.3.0",
+    "sinon": "^2.3.5",
+    "stats-webpack-plugin": "^0.6.1",
+    "webpack": "^3.0.0",
+    "webpack-dev-middleware": "^1.11.0",
+    "webpack-hot-middleware": "^2.18.0"
   }
 }
