@@ -1,29 +1,30 @@
-'use strict'
+'use strict';
 
-import React, { Component } from 'react'
-import { bindActionCreators } from 'redux'
-import { connect } from 'react-redux'
+import React, {Component} from 'react'
+import PropTypes from 'prop-types'
+import {bindActionCreators} from 'redux'
+import {connect} from 'react-redux'
 import Counter from './Counter'
 import * as CounterActions from '../actions'
 
 class CounterApp extends Component {
-  render () {
-    const { counter, dispatch } = this.props
+  render() {
+    const {counter, dispatch} = this.props;
     return (
-      <Counter value={counter} {...bindActionCreators (CounterActions, dispatch)} />
+      <Counter value={counter} {...bindActionCreators(CounterActions, dispatch)} />
     )
   }
 }
 
 CounterApp.propTypes = {
-  counter:React.PropTypes.object.isRequired,
-  dispatch:React.PropTypes.func.isRequired
-}
+  counter: PropTypes.object.isRequired,
+  dispatch: PropTypes.func.isRequired
+};
 
-function select (state) {
+function select(state) {
   return {
-    counter: state.counter.get ('value')
+    counter: state.counter.get('value')
   }
 }
 
-export default connect (select) (CounterApp)
+export default connect(select)(CounterApp)
