@@ -1,6 +1,6 @@
-'use strict'
+'use strict';
 
 import chai from 'chai'
-
 import chaiImmutable from 'chai-immutable'
-chai.use(chaiImmutable)
+
+chai.use(chaiImmutable);
